fix: change data type in offset calculator

+++ backend/src/main/java/edu/architect_711/words/service/OffsetCalculator.java
@@ -1,7 +1,7 @@
 package edu.architect_711.words.service;
 
 public class OffsetCalculator {
-    public static Long regular(final Long size, final Long page) {
+    public static int regular(final int size, final int page) {
         return size * page;
     }
 }
